% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{exampleFour}
\alias{exampleFour}
\title{Fourth Example}
\usage{
exampleFour()
}
\value{
None
}
\description{
A simple example invoking a derived Rcpp logger.
}
\details{
Note that this still triggers R warnings, so consider it work in progress.
}
