% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{exampleRsink}
\alias{exampleRsink}
\title{spdlog Example using a sink for R}
\usage{
exampleRsink()
}
\value{
None
}
\description{
A simple example invoking a derived R/Rcpp logger. Also demonstrates the
stopwatch feature. For more features see the 'spdlog' documnetation.
}
\details{
Note that this no longer triggers R warnings thanks to excellent help by
Gabi Melman.
}
\examples{
exampleRsink()
}
