% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/stopwatch.R
\name{log_setup}
\alias{log_setup}
\alias{log_filesetup}
\alias{log_drop}
\alias{log_set_pattern}
\alias{log_set_level}
\alias{log_trace}
\alias{log_debug}
\alias{log_info}
\alias{log_warn}
\alias{log_error}
\alias{log_critical}
\alias{get_stopwatch}
\alias{elapsed_stopwatch}
\alias{format_stopwatch}
\alias{print.stopwatch}
\alias{format.stopwatch}
\title{R Accessor Functions for spdlog Logger}
\usage{
log_setup(name = "default", level = "warn")

log_filesetup(filename, name = "default", level = "warn")

log_drop(name)

log_set_pattern(s)

log_set_level(s)

log_trace(s)

log_debug(s)

log_info(s)

log_warn(s)

log_error(s)

log_critical(s)

get_stopwatch()

elapsed_stopwatch(sw)

format_stopwatch(sw)

\method{print}{stopwatch}(x, ...)

\method{format}{stopwatch}(x, ...)
}
\arguments{
\item{name}{A character variable with the logging instance name, default value is \sQuote{default}.}

\item{level}{A character variable with the default logging level, default value is \sQuote{warn}.}

\item{filename}{A character variable with the logging filename if a file-based logger is
instantiated.}

\item{s}{A character variable with the logging pattern, level or message.}

\item{sw}{An S3 object of type \code{stopwatch}.}

\item{x}{An S3 object of type \code{stopwatch}.}

\item{...}{Dotted argument required by generic, unused here.}
}
\value{
Nothing is returned from these functions as they are invoked for their side-effects.
}
\description{
Several R-level functions can access the \code{spdlog} logging facilties. As \code{spdlog}
is a C++-level logging library, these are R function permit concurrent logging from both
R and C++.
}
\details{
Several functions are provided:
\describe{
   \item{\code{log_setup}}{Initializes a logger (which becomes the default logger).}
   \item{\code{log_filesetup}}{Initializes a file-based logger (which becomes the default).}
   \item{\code{log_drop}}{Removes logger (which in general should not be needed).}
   \item{\code{log_set_pattern}}{Changes the default logging message pattern.}
   \item{\code{log_set_level}}{Sets the logging level threshold.}
   \item{\code{log_trace}}{Logs a trace-level message.}
   \item{\code{log_debug}}{Logs a debug-level message.}
   \item{\code{log_info}}{Logs a info-level message.}
   \item{\code{log_warn}}{Logs a warn-level message.}
   \item{\code{log_error}}{Logs a error-level message.}
   \item{\code{log_critical}}{Logs a critical-level message.}
   \item{\code{get_stopwatch}}{Returns a stopwatch object (as an S3 object).}
   \item{\code{elapsed_stopwatch}}{Returns elapsed time for stopwatch in seconds.}
   \item{\code{format_stopwatch}}{Returns elapsed time for stopwatch as character variable.}
}
The \code{stopwatch} object has \code{print} and \code{format} methods.

Supported logging levels are, in order of increasing threshold values, \sQuote{trace},
\sQuote{debug}, \sQuote{warn}, \sQuote{info}, \sQuote{warn}, \sQuote{error}, and
\sQuote{critical}.  A message issued below the current threshold is not displayed whereas
a message at or above the current threshold is displayed.  The default level is \sQuote{warn}.
}
\examples{
log_setup("demo")
log_info("this message is NOT seen")
log_set_level("debug")
log_info("this message is seen")
log_warn("as is this message")

}
