% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{exampleFour}
\alias{exampleFour}
\title{Fourth Example}
\usage{
exampleFour()
}
\value{
None
}
\description{
A simple example invoking a derived Rcpp logger.
}
\details{
Note that this no longer triggers R warnings thanks to excellent help by
Gabi Melman.
}
\examples{
exampleFour()
}
